+++ app/controllers/exam_templates_controller.rb
@@ -13,7 +13,7 @@ class ExamTemplatesController < ApplicationController
 
   def index
     @assignment = Assignment.find(params[:assignment_id])
-    @exam_templates = @assignment.exam_templates.includes(:template_divisions)
+    @exam_templates = @assignment.exam_templates.order(:created_at).includes(:template_divisions)
   end
 
   # Creates a new instance of the exam template.